fix(session): scope user session state per request with React Context
Module-scope mutable state in lib/user-session-sync.ts was shared across
concurrent SSR requests because Node keeps module state alive between
requests. Session state now lives in a store instance created by a new
SessionProvider, so each request/component tree gets its own copy, and
getServerSnapshot returns a frozen default state object.

- Add createSessionStore() encapsulating state, listeners, localStorage
  persistence and BroadcastChannel/storage cross-tab syncing
- Add SessionProvider + context; useUserSessionSync reads the scoped store
- Tear down the channel and storage listener when the provider unmounts
- Mount SessionProvider in the root layout
